+++ src/decorators/BotDeclaration.ts
@@ -7,10 +7,10 @@ import "reflect-metadata";
 export interface IBotDeclarationSettings {
     endpoint: string;
     storage: Storage;
-    appId: string | undefined;
-    appPassword: string | undefined;
-    certificatePrivateKey?: string | undefined;
-    certificateThumbprint?: string | undefined;
+    appId: string | undefined | (() => string);
+    appPassword: string | undefined | (() => string);
+    certificatePrivateKey?: string | undefined  | (() => string);
+    certificateThumbprint?: string | undefined  | (() => string);
     namespace?: string | undefined;
 }
 
@@ -18,19 +18,19 @@ export interface IBotDeclarationSettings {
  * Decorator function for Bots
  * @param endpoint The endpoint to expose for the bot (typically '/api/messages')
  * @param storage The Storage to use for the bot (ex: new MemoryStorage())
- * @param appId The App Id for the bot
- * @param appPassword The app password for the bot
- * @param certificatePrivateKey (Optional) Client Certificate Credential - PK
- * @param certificateThumbprint (Optional) Client Certificate Credential - TP
+ * @param appId The App Id for the bot, string or function that returns a string
+ * @param appPassword The app password for the bot, string or function that returns a string
+ * @param certificatePrivateKey (Optional) Client Certificate Credential - PK, , string or function that returns a string
+ * @param certificateThumbprint (Optional) Client Certificate Credential - TP, , string or function that returns a string
  * @param namespace (Optional) Namespace to be appended to storage keys, defaults to empty string
  */
 export function BotDeclaration(
     endpoint: string,
     storage: Storage,
-    appId: string | undefined,
-    appPassword: string | undefined,
-    certificatePrivateKey?: string | undefined,
-    certificateThumbprint?: string | undefined,
+    appId: string | undefined| (() => string),
+    appPassword: string | undefined| (() => string),
+    certificatePrivateKey?: string | undefined| (() => string),
+    certificateThumbprint?: string | undefined| (() => string),
     namespace?: string | undefined) {
     // tslint:disable-next-line: ban-types
     return (target: Function) => Reflect.defineMetadata("msteams:bot", <IBotDeclarationSettings>{ endpoint, storage, appId, appPassword, certificatePrivateKey, certificateThumbprint, namespace }, target);